# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Feedme Cookbook API
  version: 1.0.0
  description: >
    REST API for the Feedme cookbook demo app. Data is stored in-memory on the
    server (no database) and resets on restart. All endpoints are served under
    the `/api` prefix. Protected endpoints require a Bearer JWT obtained from
    `/auth/login` or `/auth/register`.
servers:
  - url: http://localhost:8080/api
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /recipes:
    get:
      tags: [Recipes]
      summary: List recipes
      parameters:
        - { in: query, name: search, schema: { type: string } }
        - in: query
          name: sort
          schema:
            type: string
            enum: [popularity, views, rating, cooking-time, newest]
        - { in: query, name: maxCookingTime, schema: { type: integer } }
        - { in: query, name: mine, schema: { type: boolean } }
        - { in: query, name: limit, schema: { type: integer } }
      responses:
        "200":
          description: Recipe list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Recipe" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Recipes]
      summary: Create a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecipeInput" }
      responses:
        "201":
          description: Created recipe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /recipes/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Recipes]
      summary: Get a recipe (increments view count)
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Recipes]
      summary: Update a recipe (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecipeInput" }
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Recipes]
      summary: Delete a recipe (owner only)
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /recipes/{id}/like:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    put:
      tags: [Recipes]
      summary: Like a recipe
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Recipes]
      summary: Remove a like
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
        "404": { $ref: "#/components/responses/NotFound" }
  /recipes/{id}/comments:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Comments]
      summary: List comments for a recipe
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Comment" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Comments]
      summary: Add a comment to a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommentInput" }
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Message" }
    Forbidden:
      description: You do not own this resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Message" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Message" }
...
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
  schemas:
    Message:
      type: object
      properties:
        message: { type: string }
    ValidationError:
      type: object
      properties:
        message: { type: string, example: Validation failed }
        errors:
          type: object
          additionalProperties: { type: string }
          description: Map of field name to error message.
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        bio: { type: string }
        avatarUrl: { type: string, format: uri }
    Ingredient:
      type: object
      required: [name, amount]
      properties:
        name: { type: string }
        amount: { type: string }
...
    RecipeInput:
      type: object
      required: [title, imageUrl, description, ingredients, directions, cookingTimeMins]
      properties:
        title: { type: string, minLength: 2 }
        imageUrl: { type: string, format: uri }
        description: { type: string }
        ingredients:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/Ingredient" }
        directions: { type: string }
        cookingTimeMins: { type: integer, minimum: 1 }
    Recipe:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        imageUrl: { type: string, format: uri }
        description: { type: string }
        ingredients:
          type: array
          items: { $ref: "#/components/schemas/Ingredient" }
        directions: { type: string }
        cookingTimeMins: { type: integer }
        views: { type: integer }
        likesCount: { type: integer }
        likedByMe: { type: boolean }
        commentsCount: { type: integer }
        author: { $ref: "#/components/schemas/User" }
...
    CommentInput:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1, maxLength: 1000 }
    Comment:
      type: object
      properties:
        id: { type: string }
        targetType: { type: string, enum: [recipe, cookbook] }
        targetId: { type: string }
        text: { type: string }
        createdAt: { type: string, format: date-time }
        author: { $ref: "#/components/schemas/User" }